Start conversion to ES6 with transpiler
Many of the new libraries that support a react / flux application use
the ES6 syntax and need to be transpiled into javascript. Since I would
like to use some of them with this app I will have to use the babel
transpiler to get the code working and might as well take advantage of
some of the new syntax in my own code.