Fixed jittery panning after a fresh page load.
The tile viewer wasn't panning smoothly when the site transitioned from
the repo details page to the tile viewer page. It turns out this was due
to an extra block of code that was stopping the pan prematurely and
trying to start another one. Removing that block of code fixed the
issue and doesn't seem to have had any negative side effects.