Stops wrong data from being displayed temporarily.
If there isn't a check to make sure that the uuid matches the repo meta
information, then the page can display the old data from the previous
repo before the current one is loaded. This causes a flash of old
content to be displayed and is not good, especially if on a slow
connection.