Fixes div proliferation when switching pages.
For some reason, the OverlayTrigger component in react-bootstrap adds a
ton of divs to the page when it is used and they aren't cleaned up after
it goes out of scope. Rather than debug their code, I switched over to
just using popovers and calling the javascript code from bootstrap
directly to activate them.