Adds dist directory creation.
There is now a Grunt task that creates a dist directory that should be
used by the dvid -webclient command. This prevents all the additional
files, like LICENSE.txt and README.md, from being served up by dvid.
Only files that have been added to the copy task in the Gruntfile will
be available for dvid to serve.