Fixes scrolling issues when not at top of window.
The react router code was hijacking the scrolling behaviour to try to
emulate the way a browser always starts at the top of a page after a
transition. This was causing the page to jump around when scrolling in
the tile viewer, because it updates the url. Thankfully there is an
ignoreScrollBehaviour flag that stops this. This fixes issue #2