Fixing dvid version link to github.
As reported in #6, sometimes the dvid version might not be the exact
version string found in github. This can be caused by adding some code
to the server after checkout, which will then append '-dirty' to the
version number. This update removes all suffixes from the version number
in the link, so that it still works when trying to connect to github.
It also adds a default link to the top level of the repo if there is no
version number.