Fixes incorrect DAG rendering.
The updated properties were not getting passed into the initDag method.
It was using the old values and so it was rendering the data from the
previous repo the firest time through and then rendering it correctly
the second time through, once the correct data had been fetched from the
remote server. Fixes #11