Fixed arrowhead changes after SVG download.
The arrowheads were changing from filled to not filled after the
download button was clicked. The problem comes from the
getComputedStyles function, which seems to convert a fill of #000 to a
fill of auto with an opacity of rgb(0,0,0). This doesn't get passed back
to the constructed SVG and so the correct arrowhead is not drawn. To fix
it I changed the color from #000 to #111 and the correct fill value is
assigned to the downloaded SVG. Nobody will notice the difference
between #000 and #111 on a normal monitor, so we should be fine.