Fixes whitespace issue on the about page.
The about page was showing the github link concatenated with the next
word in the sentence "githubas".

It seems that versions of react from 0.9 onwards changed the way
whitespace is preserved, so that a tag at the end of a line will not
preserve the whitespace after it. The simple fix is to move the text so
that the parser doesn't get it wrong, but one must remember not to end a
line with a tag and start a newline with text.

eg:
--bad--

found on <a href="https://github.com/janelia-flyem/dvid">github</a>
as well as the automatically updated

--good--

found on <a href="https://github.com/janelia-flyem/dvid">github</a> as
well as the automatically updated