Log panel now updates for logs without entries.
Previously, when moving the mouse over a node that had no log entires,
the log display was not update. This caused some confusion about which
log entires belong to which node.

The update changes the log display, so that there can be no confusion
about which logs belong to the node.