Fixes repo log not updating between repos.
The repository log was displaying the values from the last selected
repository, so it was showing the wrong data unless you clicked on the
repository link twice. This was because the log store was not being
initialized with the correct properties and needed to be updated from
the componentWillUpdate method, which provides the correct properties.