Fixed missing coordinates when segmentation enabled.
When the segmentation labels were enabled on the tile viewer, the x,y,z
coordinate pop up was not working and the keyboard commands were failing
too. This was because the mouse event used to trigger them was attached
to a canvas element that was below the segmentation canvas element.
This change attached the event to both of the canvas elements.