Adds method to change focus without scrolling
The standard behavior of the focus method is to focus keyboard input on
the chosen element and scroll it into view. The scroll behavior was not
working for the instance where we select the keyboard input element to
allow keyboard control of the tile viewer as it was making the page jump
around unexpectedly. This stops the jump from happening.