Fixed x,y,z input form below the tile viewer.
This form would break the tile viewer by passing the NaN value if any of
the fields were left blank on submission. It seems crazy to have to fill
out the form completely to change just one axis, so the app now works
out the values for the fields that were left blank.