Altered rendering and caching to make image appear crisper
The image should appear to be crisper with less blurring during the
changes between layers and positions in the plane. Increasing the number
of cached layers does this to some extent. Changing OpenSeaDragon to use
immediate rendering of the tiles removes the blurring that was
occurring. The trade off is that the screen can go black, instead of
blurry.