Tiled datatypes now checked for MaxTileCoord before loading source.
This makes it possible to load in all the meta data needed to display a
tiled data type without having to call a source data type to get the
MaxPoint information. It also means that the source doesn't even have to
exist, which will be the case with many of the tile sources that are
generated from external systems.