Fixes missing tiles when using legacy Source info
If the tiles were generated from a "Source" such as a grey scale data
set, then they don't necessarily have MaxTileCoord and MinTileCoord meta
information. However, since dvid will always serve those values as an
array of zeros we have to use the "Source" attribute as a check to make
sure that we are loading the correct data.