Adds commit button to RepoDAG component.
This button will allow someone to lock a node and add a log message to
the commit. It will try and lock whichever node is selected in the DAG.
If it can not lock the node an error message will be displayed. If
successful, the log will be updated and the node will be locked.