Removes use "new repo" button text from home page
When the DVID server starts up for the first time and there are no
repositories it shows a message on the home page. This message
originally had some text explaining how to use the "New Repository"
button to add a new repository. Since this button is not enabled unless
the admin features are turned on, I replaced the text with a request to
talk to the server administrator instead.