Quick DAG hack to make code work with missing data
There were some instances where the whole application would fail if the
DAG couldn't find a piece of data it was expecting. This adds a check to
make sure the console will still load even if that data is missing.
Further exploration is needed to figure out why that data is missing.