Fixed template issue that caused a console error.
The error was "Any use of a keyed object should be wrapped in
React.addons.createFragment(object)". This was caused by the use of
double curly brackets ({{}}) instead of single ({}). Annoying bug to
find, but fixed. Also found that I was pushing an array onto an array
instead of merging the two. This has been fixed as well.