Adds settings functions and style cleanup.
There are now settings functions that can do more that just return the
values set in the settings.json. This makes it possible to decide which
host name to use if one isn't set in the settings.json file.

Also fixed a bunch of styling issues to make the site look nicer.