Fixes file loading error in Files list.
The code was checking a 'loaded' boolean, but that was getting set, even
if the request failed. The problem was that the code sent a request even
when the repo id wasn't present, failed the request and set the loaded
flag. The update waits to send the request until the repo id is present.