Removing .eslintrc file.
It seems we can add the eslintConfig parameter to our package.json file
and that is used by eslint instead. This removes a hidden dot file,
which are never good for new developers. This also switches the linter
to extend the react-app rules instead of airbnb rules. The react-app
rules do not enforce style, but do find errors. Further investigation
will be done on style enforcement with prettier.