Adds protocol option to dvid.connect calls.
By default, the dvid api will connect to a server on the standard
http(80) port. However, sometimes it is being served via the https(443)
port and needs to make it's async requests to that port instead. This
update makes sure that the correct protocol is used by default. If you
really need to connect to an http port for the data calls, you can
specify the protocol to use in the settings.json file and it will use
those instead.

eg:

{
  dvid: {
    host: <host_name>,
    port: <port_number>,
    protocol: <http|https>
  },
  ...
}