Adds conditional load for Neuroglancer component.
Support for webGL2 is not universal and if missing from your browser of
choice, will cause neuroglancer to fail to load. That load failure
happens as soon as the code is imported, which means it will take out
the entire site. So, we lazy load it in to the site and only load it if
the browser supports it.