Switching to HashRouter due to limitations.
The DVID server is not set up to serve the consoles index.html for any
urls that don't match existing paths. Instead, it serves up a 404 error.
This is probably the right thing to do, but it means that the console
needs to use the HashRouter. Doing so makes sure that if someone
refreshes the console on a repo page or anything that isn't '/' the
correct response will be sent back to the browser from dvid.