Moves server configuration into .env.development
By default the console should use the same host as it is being served
from to fetch api requests. This doesn't work in development, when the
code is not being served up by a dvid server. In these cases, are remote
dvid server is used to test out the code.